Refactor controller tests
